requirements.txt
biopython>==1.70
ete3>=3.1.2
gffutils>=0.9
pandas>=1.1.5
pronto>=2.3.2
requests>=2.22.0
